--- homeassistant/components/ua_air_raid_siren/translations/ar.json
@@ -1,23 +1,0 @@
-{
-    "config": {
-        "step": {
-            "user": {
-                "data": {
-                    "language": "\u0627\u0644\u0644\u063a\u0629",
-                    "name": "\u0627\u0633\u0645 \u0627\u0644\u062a\u0643\u0627\u0645\u0644"
-                },
-                "description": "\u0642\u0645 \u0628\u0625\u0639\u062f\u0627\u062f \u062a\u0643\u0627\u0645\u0644 OpenWeatherMap. \u0644\u0625\u0646\u0634\u0627\u0621 \u0645\u0641\u062a\u0627\u062d API \u060c \u0627\u0646\u062a\u0642\u0644 \u0625\u0644\u0649 https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u0627\u0644\u0644\u063a\u0629"
-                }
-            }
-        }
-    }
-}
