--- homeassistant/components/ua_air_raid_siren/translations/bg.json
@@ -1,34 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\u041c\u0435\u0441\u0442\u043e\u043f\u043e\u043b\u043e\u0436\u0435\u043d\u0438\u0435\u0442\u043e \u0432\u0435\u0447\u0435 \u0435 \u043a\u043e\u043d\u0444\u0438\u0433\u0443\u0440\u0438\u0440\u0430\u043d\u043e"
-        },
-        "error": {
-            "cannot_connect": "\u041d\u0435\u0443\u0441\u043f\u0435\u0445 \u043f\u0440\u0438 \u0441\u0432\u044a\u0440\u0437\u0432\u0430\u043d\u0435",
-            "invalid_api_key": "\u041d\u0435\u0432\u0430\u043b\u0438\u0434\u0435\u043d API \u043a\u043b\u044e\u0447"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API \u043a\u043b\u044e\u0447",
-                    "language": "\u0415\u0437\u0438\u043a",
-                    "latitude": "\u0413\u0435\u043e\u0433\u0440\u0430\u0444\u0441\u043a\u0430 \u0448\u0438\u0440\u0438\u043d\u0430",
-                    "longitude": "\u0413\u0435\u043e\u0433\u0440\u0430\u0444\u0441\u043a\u0430 \u0434\u044a\u043b\u0436\u0438\u043d\u0430",
-                    "mode": "\u0420\u0435\u0436\u0438\u043c",
-                    "name": "\u0418\u043c\u0435 \u043d\u0430 \u0438\u043d\u0442\u0435\u0433\u0440\u0430\u0446\u0438\u044f\u0442\u0430"
-                },
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u0415\u0437\u0438\u043a",
-                    "mode": "\u0420\u0435\u0436\u0438\u043c"
-                }
-            }
-        }
-    }
-}
