--- homeassistant/components/ua_air_raid_siren/translations/ca.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "La ubicaci\u00f3 ja est\u00e0 configurada"
-        },
-        "error": {
-            "cannot_connect": "Ha fallat la connexi\u00f3",
-            "invalid_api_key": "Clau API inv\u00e0lida"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Clau API",
-                    "language": "Idioma",
-                    "latitude": "Latitud",
-                    "longitude": "Longitud",
-                    "mode": "Mode",
-                    "name": "Nom de la integraci\u00f3"
-                },
-                "description": "Configura la integraci\u00f3 OpenWeatherMap. Per generar la clau API, ves a https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Idioma",
-                    "mode": "Mode"
-                }
-            }
-        }
-    }
-}
