--- homeassistant/components/ua_air_raid_siren/translations/de.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Standort ist bereits konfiguriert"
-        },
-        "error": {
-            "cannot_connect": "Verbindung fehlgeschlagen",
-            "invalid_api_key": "Ung\u00fcltiger API-Schl\u00fcssel"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API-Schl\u00fcssel",
-                    "language": "Sprache",
-                    "latitude": "Breitengrad",
-                    "longitude": "L\u00e4ngengrad",
-                    "mode": "Modus",
-                    "name": "Name der Integration"
-                },
-                "description": "Richte die OpenWeatherMap-Integration ein. Zum Generieren des API-Schl\u00fcssels gehe auf https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Sprache",
-                    "mode": "Modus"
-                }
-            }
-        }
-    }
-}
