--- homeassistant/components/ua_air_raid_siren/translations/en.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Location is already configured"
-        },
-        "error": {
-            "cannot_connect": "Failed to connect",
-            "invalid_api_key": "Invalid API key"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API Key",
-                    "language": "Language",
-                    "latitude": "Latitude",
-                    "longitude": "Longitude",
-                    "mode": "Mode",
-                    "name": "Name of the integration"
-                },
-                "description": "Set up OpenWeatherMap integration. To generate API key go to https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Language",
-                    "mode": "Mode"
-                }
-            }
-        }
-    }
-}
