--- homeassistant/components/ua_air_raid_siren/translations/et.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Nende koordinaatidele on OpenWeatherMapi sidumine  juba tehtud."
-        },
-        "error": {
-            "cannot_connect": "\u00dchendamine nurjus",
-            "invalid_api_key": "Vigane API v\u00f5ti"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "OpenWeatherMapi API v\u00f5ti",
-                    "language": "Keel",
-                    "latitude": "Laiuskraad",
-                    "longitude": "Pikkuskraad",
-                    "mode": "Re\u017eiim",
-                    "name": "Sidumise nimi"
-                },
-                "description": "Seadista OpenWeatherMapi sidumine. API-v\u00f5tme loomiseks mine aadressile https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Keel",
-                    "mode": "Re\u017eiim"
-                }
-            }
-        }
-    }
-}
