--- homeassistant/components/ua_air_raid_siren/translations/fr.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "L'emplacement est d\u00e9j\u00e0 configur\u00e9"
-        },
-        "error": {
-            "cannot_connect": "\u00c9chec de connexion",
-            "invalid_api_key": "Cl\u00e9 d'API non valide"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Cl\u00e9 d'API",
-                    "language": "Langue",
-                    "latitude": "Latitude",
-                    "longitude": "Longitude",
-                    "mode": "Mode",
-                    "name": "Nom de l'int\u00e9gration"
-                },
-                "description": "Configurez l'int\u00e9gration OpenWeatherMap. Pour g\u00e9n\u00e9rer la cl\u00e9 API, acc\u00e9dez \u00e0 https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Langue",
-                    "mode": "Mode"
-                }
-            }
-        }
-    }
-}
