--- homeassistant/components/ua_air_raid_siren/translations/he.json
@@ -1,34 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\u05ea\u05e6\u05d5\u05e8\u05ea \u05d4\u05de\u05d9\u05e7\u05d5\u05dd \u05db\u05d1\u05e8 \u05e0\u05e7\u05d1\u05e2\u05d4"
-        },
-        "error": {
-            "cannot_connect": "\u05d4\u05d4\u05ea\u05d7\u05d1\u05e8\u05d5\u05ea \u05e0\u05db\u05e9\u05dc\u05d4",
-            "invalid_api_key": "\u05de\u05e4\u05ea\u05d7 API \u05dc\u05d0 \u05d7\u05d5\u05e7\u05d9"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "\u05de\u05e4\u05ea\u05d7 API",
-                    "language": "\u05e9\u05e4\u05d4",
-                    "latitude": "\u05e7\u05d5 \u05e8\u05d5\u05d7\u05d1",
-                    "longitude": "\u05e7\u05d5 \u05d0\u05d5\u05e8\u05da",
-                    "mode": "\u05de\u05e6\u05d1",
-                    "name": "\u05e9\u05dd \u05d4\u05e9\u05d9\u05dc\u05d5\u05d1"
-                },
-                "title": "\u05de\u05e4\u05ea OpenWeather"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u05e9\u05e4\u05d4",
-                    "mode": "\u05de\u05e6\u05d1"
-                }
-            }
-        }
-    }
-}
