--- homeassistant/components/ua_air_raid_siren/translations/hu.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "A hely m\u00e1r konfigur\u00e1lva van"
-        },
-        "error": {
-            "cannot_connect": "Sikertelen csatlakoz\u00e1s",
-            "invalid_api_key": "\u00c9rv\u00e9nytelen API kulcs"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API kulcs",
-                    "language": "Nyelv",
-                    "latitude": "Sz\u00e9less\u00e9g",
-                    "longitude": "Hossz\u00fas\u00e1g",
-                    "mode": "M\u00f3d",
-                    "name": "Az integr\u00e1ci\u00f3 neve"
-                },
-                "description": "Az OpenWeatherMap integr\u00e1ci\u00f3 be\u00e1ll\u00edt\u00e1sa. Az API kulcs l\u00e9trehoz\u00e1s\u00e1hoz l\u00e1togasson el a https://openweathermap.org/appid oldalra",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Nyelv",
-                    "mode": "M\u00f3d"
-                }
-            }
-        }
-    }
-}
