--- homeassistant/components/ua_air_raid_siren/translations/id.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Lokasi sudah dikonfigurasi"
-        },
-        "error": {
-            "cannot_connect": "Gagal terhubung",
-            "invalid_api_key": "Kunci API tidak valid"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Kunci API",
-                    "language": "Bahasa",
-                    "latitude": "Lintang",
-                    "longitude": "Bujur",
-                    "mode": "Mode",
-                    "name": "Nama integrasi"
-                },
-                "description": "Siapkan integrasi OpenWeatherMap. Untuk membuat kunci API, buka https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Bahasa",
-                    "mode": "Mode"
-                }
-            }
-        }
-    }
-}
