--- homeassistant/components/ua_air_raid_siren/translations/it.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "La posizione \u00e8 gi\u00e0 configurata"
-        },
-        "error": {
-            "cannot_connect": "Impossibile connettersi",
-            "invalid_api_key": "Chiave API non valida"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Chiave API",
-                    "language": "Lingua",
-                    "latitude": "Latitudine",
-                    "longitude": "Logitudine",
-                    "mode": "Modalit\u00e0",
-                    "name": "Nome dell'integrazione"
-                },
-                "description": "Configura l'integrazione di OpenWeatherMap. Per generare la chiave API, vai su https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Lingua",
-                    "mode": "Modalit\u00e0"
-                }
-            }
-        }
-    }
-}
