--- homeassistant/components/ua_air_raid_siren/translations/ja.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\u30ed\u30b1\u30fc\u30b7\u30e7\u30f3\u306f\u3059\u3067\u306b\u8a2d\u5b9a\u3055\u308c\u3066\u3044\u307e\u3059"
-        },
-        "error": {
-            "cannot_connect": "\u63a5\u7d9a\u306b\u5931\u6557\u3057\u307e\u3057\u305f",
-            "invalid_api_key": "\u7121\u52b9\u306aAPI\u30ad\u30fc"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API\u30ad\u30fc",
-                    "language": "\u8a00\u8a9e",
-                    "latitude": "\u7def\u5ea6",
-                    "longitude": "\u7d4c\u5ea6",
-                    "mode": "\u30e2\u30fc\u30c9",
-                    "name": "\u30a4\u30f3\u30c6\u30b0\u30ec\u30fc\u30b7\u30e7\u30f3\u306e\u540d\u524d"
-                },
-                "description": "OpenWeatherMap\u306e\u30a4\u30f3\u30c6\u30b0\u30ec\u30fc\u30b7\u30e7\u30f3\u3092\u30bb\u30c3\u30c8\u30a2\u30c3\u30d7\u3057\u307e\u3059\u3002API\u30ad\u30fc\u3092\u751f\u6210\u3059\u308b\u306b\u306f\u3001https://openweathermap.org/appid \u306b\u30a2\u30af\u30bb\u30b9\u3057\u3066\u304f\u3060\u3055\u3044",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u8a00\u8a9e",
-                    "mode": "\u30e2\u30fc\u30c9"
-                }
-            }
-        }
-    }
-}
