--- homeassistant/components/ua_air_raid_siren/translations/ko.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\uc704\uce58\uac00 \uc774\ubbf8 \uad6c\uc131\ub418\uc5c8\uc2b5\ub2c8\ub2e4"
-        },
-        "error": {
-            "cannot_connect": "\uc5f0\uacb0\ud558\uc9c0 \ubabb\ud588\uc2b5\ub2c8\ub2e4",
-            "invalid_api_key": "API \ud0a4\uac00 \uc798\ubabb\ub418\uc5c8\uc2b5\ub2c8\ub2e4"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API \ud0a4",
-                    "language": "\uc5b8\uc5b4",
-                    "latitude": "\uc704\ub3c4",
-                    "longitude": "\uacbd\ub3c4",
-                    "mode": "\ubaa8\ub4dc",
-                    "name": "\ud1b5\ud569 \uad6c\uc131\uc694\uc18c \uc774\ub984"
-                },
-                "description": "OpenWeatherMap \ud1b5\ud569 \uad6c\uc131\uc694\uc18c\ub97c \uc124\uc815\ud569\ub2c8\ub2e4. API \ud0a4\ub97c \uc0dd\uc131\ud558\ub824\uba74 https://openweathermap.org/appid \ub85c \uc774\ub3d9\ud574\uc8fc\uc138\uc694",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\uc5b8\uc5b4",
-                    "mode": "\ubaa8\ub4dc"
-                }
-            }
-        }
-    }
-}
