--- homeassistant/components/ua_air_raid_siren/translations/lb.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Standuert ass scho konfigur\u00e9iert"
-        },
-        "error": {
-            "cannot_connect": "Feeler beim verbannen",
-            "invalid_api_key": "Ong\u00ebltegen API Schl\u00ebssel"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API Schl\u00ebssel",
-                    "language": "Sproch",
-                    "latitude": "Breedegrad",
-                    "longitude": "L\u00e4ngegrad",
-                    "mode": "Modus",
-                    "name": "Numm vun der Integratioun"
-                },
-                "description": "OpenWeatherMap Integratioun ariichten. Fir een API Schl\u00ebssel z'erstelle g\u00e9i op https://openweathermap.org/appid",
-                "title": "OpenWeatherMap API Schl\u00ebssel"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Sproch",
-                    "mode": "Modus"
-                }
-            }
-        }
-    }
-}
