--- homeassistant/components/ua_air_raid_siren/translations/nb.json
@@ -1,24 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "OpenWeatherMap-integrering for disse koordinatene er allerede konfigurert."
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "mode": "Modus"
-                }
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Spr\u00e5k",
-                    "mode": "Modus"
-                }
-            }
-        }
-    }
-}
