--- homeassistant/components/ua_air_raid_siren/translations/pl.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Lokalizacja jest ju\u017c skonfigurowana"
-        },
-        "error": {
-            "cannot_connect": "Nie mo\u017cna nawi\u0105za\u0107 po\u0142\u0105czenia",
-            "invalid_api_key": "Nieprawid\u0142owy klucz API"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Klucz API",
-                    "language": "J\u0119zyk",
-                    "latitude": "Szeroko\u015b\u0107 geograficzna",
-                    "longitude": "D\u0142ugo\u015b\u0107 geograficzna",
-                    "mode": "Tryb",
-                    "name": "Nazwa integracji"
-                },
-                "description": "Konfiguracja integracji OpenWeatherMap. Aby wygenerowa\u0107 klucz API, przejd\u017a do https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "J\u0119zyk",
-                    "mode": "Tryb"
-                }
-            }
-        }
-    }
-}
