--- homeassistant/components/ua_air_raid_siren/translations/pt-BR.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Localiza\u00e7\u00e3o j\u00e1 est\u00e1 configurada"
-        },
-        "error": {
-            "cannot_connect": "Falha ao conectar",
-            "invalid_api_key": "Chave de API inv\u00e1lida"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Chave da API",
-                    "language": "Idioma",
-                    "latitude": "Latitude",
-                    "longitude": "Longitude",
-                    "mode": "Modo",
-                    "name": "Nome da integra\u00e7\u00e3o"
-                },
-                "description": "Configure a integra\u00e7\u00e3o do OpenWeatherMap. Para gerar a chave de API, acesse https://openweathermap.org/appid",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Idioma",
-                    "mode": "Modo"
-                }
-            }
-        }
-    }
-}
