--- homeassistant/components/ua_air_raid_siren/translations/pt.json
@@ -1,33 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "A localiza\u00e7\u00e3o j\u00e1 est\u00e1 configurada"
-        },
-        "error": {
-            "cannot_connect": "Falha na liga\u00e7\u00e3o",
-            "invalid_api_key": "Chave de API inv\u00e1lida"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "Chave da API",
-                    "language": "Idioma",
-                    "latitude": "Latitude",
-                    "longitude": "Longitude",
-                    "mode": "Modo",
-                    "name": "Nome da integra\u00e7\u00e3o"
-                }
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Idioma",
-                    "mode": "Modo"
-                }
-            }
-        }
-    }
-}
