--- homeassistant/components/ua_air_raid_siren/translations/ru.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\u041d\u0430\u0441\u0442\u0440\u043e\u0439\u043a\u0430 \u0434\u043b\u044f \u044d\u0442\u043e\u0433\u043e \u043c\u0435\u0441\u0442\u043e\u043f\u043e\u043b\u043e\u0436\u0435\u043d\u0438\u044f \u0443\u0436\u0435 \u0432\u044b\u043f\u043e\u043b\u043d\u0435\u043d\u0430."
-        },
-        "error": {
-            "cannot_connect": "\u041d\u0435 \u0443\u0434\u0430\u043b\u043e\u0441\u044c \u043f\u043e\u0434\u043a\u043b\u044e\u0447\u0438\u0442\u044c\u0441\u044f.",
-            "invalid_api_key": "\u041d\u0435\u0432\u0435\u0440\u043d\u044b\u0439 \u043a\u043b\u044e\u0447 API."
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "\u041a\u043b\u044e\u0447 API",
-                    "language": "\u042f\u0437\u044b\u043a",
-                    "latitude": "\u0428\u0438\u0440\u043e\u0442\u0430",
-                    "longitude": "\u0414\u043e\u043b\u0433\u043e\u0442\u0430",
-                    "mode": "\u0420\u0435\u0436\u0438\u043c",
-                    "name": "\u041d\u0430\u0437\u0432\u0430\u043d\u0438\u0435"
-                },
-                "description": "\u041d\u0430\u0441\u0442\u0440\u043e\u0439\u043a\u0430 Home Assistant \u0434\u043b\u044f \u0438\u043d\u0442\u0435\u0433\u0440\u0430\u0446\u0438\u0438 \u0441 OpenWeatherMap. \u0414\u043b\u044f \u043f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u044f \u043a\u043b\u044e\u0447\u0430 API, \u043f\u0435\u0440\u0435\u0439\u0434\u0438\u0442\u0435 \u043d\u0430 https://openweathermap.org/appid.",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u042f\u0437\u044b\u043a",
-                    "mode": "\u0420\u0435\u0436\u0438\u043c"
-                }
-            }
-        }
-    }
-}
