--- homeassistant/components/ua_air_raid_siren/translations/sk.json
@@ -1,19 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Umiestnenie u\u017e je nakonfigurovan\u00e9"
-        },
-        "error": {
-            "invalid_api_key": "Neplatn\u00fd API k\u013e\u00fa\u010d"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API k\u013e\u00fa\u010d",
-                    "latitude": "Zemepisn\u00e1 \u0161\u00edrka",
-                    "longitude": "Zemepisn\u00e1 d\u013a\u017eka"
-                }
-            }
-        }
-    }
-}
