--- homeassistant/components/ua_air_raid_siren/translations/sl.json
@@ -1,20 +1,0 @@
-{
-    "config": {
-        "step": {
-            "user": {
-                "data": {
-                    "mode": "Na\u010din"
-                }
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "mode": "Na\u010din"
-                }
-            }
-        }
-    }
-}
