--- homeassistant/components/ua_air_raid_siren/translations/tr.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "Konum zaten yap\u0131land\u0131r\u0131lm\u0131\u015f"
-        },
-        "error": {
-            "cannot_connect": "Ba\u011flanma hatas\u0131",
-            "invalid_api_key": "Ge\u00e7ersiz API anahtar\u0131"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API Anahtar\u0131",
-                    "language": "Dil",
-                    "latitude": "Enlem",
-                    "longitude": "Boylam",
-                    "mode": "Mod",
-                    "name": "Cihaz\u0131n ad\u0131"
-                },
-                "description": "OpenWeatherMap entegrasyonunu ayarlay\u0131n. API anahtar\u0131 olu\u015fturmak i\u00e7in https://openweathermap.org/appid adresine gidin.",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "Dil",
-                    "mode": "Mod"
-                }
-            }
-        }
-    }
-}
