--- homeassistant/components/ua_air_raid_siren/translations/uk.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\u041d\u0430\u043b\u0430\u0448\u0442\u0443\u0432\u0430\u043d\u043d\u044f \u0434\u043b\u044f \u0446\u044c\u043e\u0433\u043e \u043c\u0456\u0441\u0446\u0435\u0437\u043d\u0430\u0445\u043e\u0434\u0436\u0435\u043d\u043d\u044f \u0432\u0436\u0435 \u0432\u0438\u043a\u043e\u043d\u0430\u043d\u0435."
-        },
-        "error": {
-            "cannot_connect": "\u041d\u0435 \u0432\u0434\u0430\u043b\u043e\u0441\u044f \u043f\u0456\u0434'\u0454\u0434\u043d\u0430\u0442\u0438\u0441\u044f",
-            "invalid_api_key": "\u0425\u0438\u0431\u043d\u0438\u0439 \u043a\u043b\u044e\u0447 API"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "\u041a\u043b\u044e\u0447 API",
-                    "language": "\u041c\u043e\u0432\u0430",
-                    "latitude": "\u0428\u0438\u0440\u043e\u0442\u0430",
-                    "longitude": "\u0414\u043e\u0432\u0433\u043e\u0442\u0430",
-                    "mode": "\u0420\u0435\u0436\u0438\u043c",
-                    "name": "\u041d\u0430\u0437\u0432\u0430"
-                },
-                "description": "\u041d\u0430\u043b\u0430\u0448\u0442\u0443\u0439\u0442\u0435 Home Assistant \u0434\u043b\u044f \u0456\u043d\u0442\u0435\u0433\u0440\u0430\u0446\u0456\u0457 \u0437 OpenWeatherMap. \u0414\u043b\u044f \u043e\u0442\u0440\u0438\u043c\u0430\u043d\u043d\u044f \u043a\u043b\u044e\u0447\u0430 API, \u043f\u0435\u0440\u0435\u0439\u0434\u0456\u0442\u044c \u043d\u0430 https://openweathermap.org/appid.",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u041c\u043e\u0432\u0430",
-                    "mode": "\u0420\u0435\u0436\u0438\u043c"
-                }
-            }
-        }
-    }
-}
