--- homeassistant/components/ua_air_raid_siren/translations/zh-Hant.json
@@ -1,35 +1,0 @@
-{
-    "config": {
-        "abort": {
-            "already_configured": "\u5ea7\u6a19\u5df2\u7d93\u8a2d\u5b9a\u5b8c\u6210"
-        },
-        "error": {
-            "cannot_connect": "\u9023\u7dda\u5931\u6557",
-            "invalid_api_key": "API \u91d1\u9470\u7121\u6548"
-        },
-        "step": {
-            "user": {
-                "data": {
-                    "api_key": "API \u91d1\u9470",
-                    "language": "\u8a9e\u8a00",
-                    "latitude": "\u7def\u5ea6",
-                    "longitude": "\u7d93\u5ea6",
-                    "mode": "\u6a21\u5f0f",
-                    "name": "\u6574\u5408\u540d\u7a31"
-                },
-                "description": "\u6b32\u8a2d\u5b9a OpenWeatherMap \u6574\u5408\u3002\u8acb\u81f3 https://openweathermap.org/appid \u7522\u751f API \u91d1\u9470",
-                "title": "OpenWeatherMap"
-            }
-        }
-    },
-    "options": {
-        "step": {
-            "init": {
-                "data": {
-                    "language": "\u8a9e\u8a00",
-                    "mode": "\u6a21\u5f0f"
-                }
-            }
-        }
-    }
-}
