+++ homeassistant/components/ukraine_alarm/translations/en.json
@@ -1,0 +1,13 @@
+{
+    "config": {
+        "step": {
+            "user": {
+                "data": {
+                    "region": "Region"
+                },
+                "description": "Set up Ukraine Alarm integration. To generate API key go to https://api.ukrainealarm.com/",
+                "title": "Ukraine Alarm"
+            }
+        }
+    }
+}
