+++ requirements_all.txt
@@ -2335,7 +2335,7 @@ twitchAPI==2.5.2
 # homeassistant.components.rainforest_eagle
 uEagle==0.0.2
 
-# homeassistant.components.ua_air_raid_siren
+# homeassistant.components.ukraine_alarm
 ukrainealarm==0.0.1
 
 # homeassistant.components.unifiprotect
